+++ CHANGELOG.md
@@ -5,6 +5,7 @@ GAMSPy 1.4.0
 ------------
 - General
   - Resolve static code analysis issues to improve code quality.
+  - Return the value as a float if the given domain sets are all literals.
 - Testing
   - Set COVERAGE_CORE to sysmon to make use of the new sys.monitoring package in Python.
 - Documentation